Verify save
