+++ .gitignore
@@ -104,7 +104,7 @@ venv.bak/
 .mypy_cache/
 
 # Save File
-*.xml
+save_file.json
 
 # Log file
 log_file.txt
